chore: add public entity
